Chore [Worker] deprecate handleTaskError function
- [+] chore(worker): deprecate handleTaskError function

Note: The handleTaskError function in the worker package has been deprecated. It was used to evaluate errors encountered during task execution and determine if a retry is appropriate. However, it has been replaced with a better retry policy that reduces complexity and frees up resources for goroutines. As a result, the handleTaskError function is no longer used and has been marked as deprecated.
